split up auth
