+++ functions/src/index.ts
@@ -1,23 +1,19 @@
 /**
- * Import function triggers from their respective submodules:
- *
- * import {onCall} from "firebase-functions/v2/https";
- * import {onDocumentWritten} from "firebase-functions/v2/firestore";
- *
- * See a full list of supported triggers at https://firebase.google.com/docs/functions
+ * Write functions in their respective files, then the must be exported here to be deployed to Firebase
  */
 
-
-import { createAccount, resetPassword, beforeCreate, onUserSignup, beforeSignIn, onUserDelete, getUserProfile } from "./triggers/auth";
-import { purgeUnverifiedUsers, purgeExpiredEmails } from "./triggers/cron";
+import { createAccount, resetPassword, getUserProfile } from "./callable/auth";
+import { beforeCreate, onUserSignup, beforeSignIn, onUserDelete } from "./triggers/auth";
 import { saveCourse, getAllCourses, getCourseInfo, courseEnroll, startCourse } from "./callable/courses";
 import { saveQuiz, getQuizResponses, startQuiz, submitQuiz } from "./callable/quizzes";
 import { getCourseReports, getUserReports } from "./callable/reports";
+import { purgeUnverifiedUsers, purgeExpiredEmails } from "./triggers/cron";
 
 export {
-    createAccount, resetPassword, beforeCreate, onUserSignup, beforeSignIn, onUserDelete, getUserProfile,
-    purgeUnverifiedUsers, purgeExpiredEmails,
+    createAccount, resetPassword, getUserProfile,
+    beforeCreate, onUserSignup, beforeSignIn, onUserDelete,
     saveCourse, getAllCourses, getCourseInfo, courseEnroll, startCourse,
     saveQuiz, getQuizResponses, startQuiz, submitQuiz,
-    getCourseReports, getUserReports
+    getCourseReports, getUserReports,
+    purgeUnverifiedUsers, purgeExpiredEmails
 };
